# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Technical Test ",
        "description": "documentation for API Technical Test",
        "contact": {
            "email": "maszahid1@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/technical-test/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/transactions/mutation": {
            "get": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Endpoint ini untuk laporan mutasi dalam periode waktu maks 30 hari",
                "operationId": "52eaf6b7806f7866d1f192c767f59654",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "tanggal awal periode mutasi",
                        "schema": {
                            "type": "date"
                        },
                        "example": "2022-02-05"
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "batas periode mutasi maks 30 hari",
                        "schema": {
                            "type": "date"
                        },
                        "example": "2022-03-07"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Akan memberikan data transaksi sesuai periode yang ditentukan"
                    },
                    "422": {
                        "description": "Some variable is required"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/transactions/topup": {
            "post": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Endpoint ini untuk melakukan transaksi pengisian saldo rekening",
                "operationId": "e2ae805072d21625b026f38dde8f36e3",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "description": "deskripsi transaksi",
                                        "type": "string",
                                        "example": "topup dari OVO"
                                    },
                                    "amount": {
                                        "description": "jumlah topup minimal Rp20.000",
                                        "type": "integer",
                                        "example": "30000"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "description": "topup dari OVO",
                                    "amount": "50000"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Akan memberikan pesan berhasil melakukan transaksi"
                    },
                    "422": {
                        "description": "Some variable is required"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/transactions/withdraw": {
            "post": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Endpoint ini untuk melakukan transaksi penarikan/withdraw",
                "operationId": "24142de781670e7a8981ec8cd516233b",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "description": "deskripsi transaksi",
                                        "type": "string",
                                        "example": "tarik tunai"
                                    },
                                    "amount": {
                                        "description": "jumlah penarikan minimal dan kelipatan Rp50.000",
                                        "type": "integer",
                                        "example": "30000"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "description": "tarik tunai",
                                    "amount": "100000"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Akan memberikan pesan berhasil melakukan transaksi"
                    },
                    "422": {
                        "description": "Some variable is required"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/transactions/transfer": {
            "post": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Endpoint ini untuk melakukan transaksi transfer ke rekening yang terdafar",
                "operationId": "495c14074088c5e340e483060a408d49",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "description": "deskripsi transaksi",
                                        "type": "string",
                                        "example": "beli sepatu shopee"
                                    },
                                    "amount": {
                                        "description": "jumlah transfer minimal Rp20.000",
                                        "type": "integer",
                                        "example": "30000"
                                    },
                                    "rekening": {
                                        "description": "rekening tujuan yang terdafar",
                                        "type": "integer",
                                        "example": "57998977"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "description": "beli sepatu shopee",
                                    "amount": "100000",
                                    "rekening": 57998977
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Akan memberikan pesan berhasil melakukan transaksi"
                    },
                    "422": {
                        "description": "Some variable is required"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}